% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ilike_model.R
\name{compile}
\alias{compile}
\title{Parse .ilike file to give ilike model.}
\usage{
compile(
  filenames,
  model_parameter_list = list(),
  R_functions = FALSE,
  external_packages = c(),
  julia_bin_dir = "",
  julia_required_libraries = c(),
  verify_cpp_function_types = FALSE,
  keep_temporary_model_code = FALSE,
  nesting_level = 1,
  print_block_ends = FALSE
)
}
\arguments{
\item{filenames}{The name (and path) of the .ilike files containing the model, stored in a vector if there is more than one.}

\item{model_parameter_list}{(optional) A list containing parameters for the model.}

\item{R_functions}{(optional) If TRUE, returns R functions. If FALSE, returns C++ functions.}

\item{external_packages}{(optional) A vector of names of other R packages the functions rely on.}

\item{julia_bin_dir}{(optional) The directory containing the Julia bin file - only needed if Julia functions are used.}

\item{julia_required_libraries}{(optional) Vector of strings, each of which is a Julia package that will be installed and loaded.}

\item{verify_cpp_function_types}{(optional) If TRUE, check the types of the parameters of user-defined .cpp functions. If FALSE (default), types are not checked.}

\item{keep_temporary_model_code}{(optional) If FALSE (default), the .cpp file generated for compilation is deleted. If TRUE, this file is left in the working directory.}

\item{nesting_level}{(optional) The level of nesting of the current call to compile. A user should always use the default of 1.}

\item{print_block_ends}{(optional) If TRUE, print the end of each block of code. If FALSE (default), do not print the end of each block of code.}
}
\value{
A list containing the model details.
}
\description{
Parse .ilike file to give ilike model.
}
