% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{constant_velocity_x}
\alias{constant_velocity_x}
\title{10000 points simulated from a constant velocity linear-Gaussian dynamic model (process noise standard deviation = 1).}
\format{
A 10000*2 matrix.
}
\usage{
constant_velocity_x
}
\description{
10000 points simulated from a constant velocity linear-Gaussian dynamic model (process noise standard deviation = 1).
}
\keyword{datasets}
