% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{constant_velocity_y}
\alias{constant_velocity_y}
\title{10000 noisy observations (Gaussian noise, standard deviation 1) of 10000 points simulated from a constant velocity linear-Gaussian dynamic model (process noise standard deviation = 1).}
\format{
A 10000*1 matrix.
}
\usage{
constant_velocity_y
}
\description{
10000 noisy observations (Gaussian noise, standard deviation 1) of 10000 points simulated from a constant velocity linear-Gaussian dynamic model (process noise standard deviation = 1).
}
\keyword{datasets}
