% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampler.R
\name{importance_sample_cpp}
\alias{importance_sample_cpp}
\title{Importance sampler using cpp functions.}
\usage{
importance_sample_cpp(
  number_of_points,
  model,
  algorithm = list(),
  max_vector_size = 4e+09
)
}
\arguments{
\item{number_of_points}{The number of importance points.}

\item{model}{A model.}

\item{algorithm}{Algorithm details.}
}
\value{
Importance points.
}
\description{
Importance sampler using cpp functions.
}
\examples{
importance_sample_cpp(10000, my_model ,my_algorithm)
}
