% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportanceSampler.R
\name{ImportanceSampler}
\alias{ImportanceSampler}
\title{Importance sampler.}
\usage{
ImportanceSampler(NumberOfPoints, Model, Algorithm, MaxVectorSize = 1e+32)
}
\arguments{
\item{NumberOfPoints}{The number of importance points.}

\item{Model}{A model.}

\item{Algorithm}{Algorithm details.}
}
\value{
Importance points.
}
\description{
Importance sampler.
}
\examples{
ImportanceSampler(10000,MyModel,MyAlgorithm)
}
