% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output.R
\name{load_enk_output}
\alias{load_enk_output}
\title{Loading ensemble Kalman output into R memory.}
\usage{
load_enk_output(
  results_directory,
  ggsmc = TRUE,
  external_log_weights = c(0),
  external_target_parameters = "",
  nesting_level = 0,
  factor = 0
)
}
\arguments{
\item{results_directory}{The folder in which the results are stored.}

\item{ggsmc}{(optional) Output in tidy format for plotting in ggsmc package.}

\item{external_log_weights}{(optional; for nested output only) The weights of the importance points external to the current folder. (default is 1, to be used at the top level of nested output)}

\item{external_target_parameters}{(optional; for nested output only) The parameters of the target external to the current folder. (default is "", corresponding to no parameters)}

\item{nesting_level}{(optional; for nested output only) The level of nesting at which to extract points. (default is 0, representing the top level of nested output)}

\item{factor}{(optional; for nested output only) The factor from which to extract points. (default is 0)}
}
\value{
A list containing the ensemble members (called particles).
}
\description{
Loading ensemble Kalman output into R memory.
}
