% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output.R
\name{load_mcmc_output}
\alias{load_mcmc_output}
\title{Loading MCMC output into R memory.}
\usage{
load_mcmc_output(results_directory, ggmcmc = NULL, ilike.output = NULL)
}
\arguments{
\item{results_directory}{The folder in which the results are stored.}

\item{ggmcmc}{(optional) Output in tidy format for plotting in ggmcmc package (default is FALSE if ilike.output is set to TRUE or not set and TRUE otherwise).}

\item{ilike.output}{(optional) Output can be processed by ilike,output package (default is TRUE if ggmcmc is set to FALSE or is not set and FALSE otherwise). Choosing TRUE for this argument is incompatible with ggmcmc=TRUE, since the two packages require different formatting of the output.}
}
\value{
A list containing the MCMC chains.
}
\description{
Loading MCMC output into R memory.
}
