% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output.R
\name{load_smc_output}
\alias{load_smc_output}
\title{Loading SMC output into R memory.}
\usage{
load_smc_output(
  results_directory,
  ggsmc = TRUE,
  as.mcmc = FALSE,
  as.enk = FALSE,
  which.targets = NULL
)
}
\arguments{
\item{results_directory}{The folder in which the results are stored.}

\item{ggsmc}{(optional) Output in tidy format for plotting in gggsmc package.}

\item{as.mcmc}{(optional) Output treats particles as different MCMC chains.}

\item{as.enk}{(optional) Output treats particles as an ensemble.}

\item{which.targets}{(optional) The indices of the targets to output (defaults to all).}
}
\value{
A list containing the SMC output.
}
\description{
Loading SMC output into R memory.
}
