% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output.R
\name{pivot_wider_smc_or_enk_output}
\alias{pivot_wider_smc_or_enk_output}
\title{Take output from the functions load_smc_output or load_enk_output and convert it to a more standard format (non-tidy format with one column per dimension).}
\usage{
pivot_wider_smc_or_enk_output(output, variables = NULL)
}
\arguments{
\item{output}{Output from the functions load_smc_output or load_enk_output.}

\item{variables}{(optional) Variables to include in the output (default is all).}
}
\description{
Take output from the functions load_smc_output or load_enk_output and convert it to a more standard format (non-tidy format with one column per dimension).
}
