% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdtsc_seed}
\alias{rdtsc_seed}
\title{Gets a seed for random number generation by using the processor time stamp (calling rdtsc in C).}
\usage{
rdtsc_seed()
}
\value{
An integer.
}
\description{
Gets a seed for random number generation by using the processor time stamp (calling rdtsc in C).
}
