% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_model_rwm_ilike}
\alias{toy_model_rwm_ilike}
\title{100000 points simulated from a two-dimensional posterior using MCMC, in the format required by the ilike.output and ggsmc packages.}
\format{
A 200000 row data frame.
}
\usage{
toy_model_rwm_ilike
}
\description{
100000 points simulated from a two-dimensional posterior using MCMC, in the format required by the ilike.output and ggsmc packages.
}
\keyword{datasets}
