% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enkmfds.R
\name{enkmfds}
\alias{enkmfds}
\title{Ensemble Kalman mean-field dynamical system}
\usage{
enkmfds(
  model,
  number_of_ensemble_members = 10,
  Delta_t = 0.5,
  initial_values = list(),
  parallel = FALSE,
  results_directory = getwd(),
  enki_iterations_to_store = 2,
  write_to_file_at_each_iteration = TRUE,
  model_parameter_list = list(),
  algorithm_parameter_list = list(),
  seed = NULL,
  grain_size = 1e+05
)
}
\arguments{
\item{model}{A file containing the model, or a pre-compiled model list.}

\item{number_of_ensemble_members}{(optional) The number of ensemble members (defaults to 10).}

\item{Delta_t}{(optional) The Delta_t parameters (defaults to 0.5)}

\item{initial_values}{(optional) A list of lists containing the initial values for the chains.}

\item{parallel}{(optional) Set to true to perform the importance sampling in parallel, false for serial.}

\item{results_directory}{The name of the directory to which results will be written.}

\item{enki_iterations_to_store}{(optional) The number of iterations of EnK output stored in memory as the algorithm is running (cannot be fewer than 2).}

\item{write_to_file_at_each_iteration}{(optional) Do we write the algorithm output to file at each EnK step (TRUE/FALSE)?}

\item{model_parameter_list}{(optional) A list containing parameters for the model.}

\item{algorithm_parameter_list}{(optional) A list containing named parameters for the algorithm.}

\item{seed}{(optional) The seed for the random number generator.}

\item{grain_size}{(optional) Sets a minimum chunk size for parallelisation (see https://oneapi-src.github.io/oneTBB/main/tbb_userguide/Controlling_Chunking_os.html).}
}
\value{
Nothing: output can be found in the output_directory.
}
\description{
Ensemble Kalman mean-field dynamical system
}
