% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaussian}
\alias{gaussian}
\title{100 points simulated from a Gaussian distribution.}
\format{
A vector of 100 variables
}
\usage{
gaussian
}
\description{
Generated using rnorm(100).
}
\keyword{datasets}
