% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilike-package.R
\docType{package}
\name{ilike}
\alias{ilike}
\title{ilike: Methods for Bayesian computation when the likelihood is intractable.}
\description{
It contains some really useful stuff.
}
\section{Functions}{

Some functions.
}

