% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampler.R
\name{importance_sample}
\alias{importance_sample}
\title{Importance sampler.}
\usage{
importance_sample(model, algorithm, messages = FALSE)
}
\arguments{
\item{model}{A model.}

\item{algorithm}{algorithm details.}

\item{messages}{Set to TRUE to have messages printed to console.}
}
\value{
Importance points.
}
\description{
Importance sampler.
}
