% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampler.R
\name{importance_sample}
\alias{importance_sample}
\title{Importance sampler.}
\usage{
importance_sample(
  number_of_points,
  model,
  algorithm = list(),
  max_vector_size = 4e+09
)
}
\arguments{
\item{number_of_points}{The number of importance points.}

\item{model}{A model.}

\item{algorithm}{algorithm details.}
}
\value{
Importance points.
}
\description{
Importance sampler.
}
\examples{
importance_sample(10000,my_model,my_algorithm)
}
