% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampler.R
\name{importance_sample}
\alias{importance_sample}
\title{Importance sampler.}
\usage{
importance_sample(
  model,
  number_of_importance_points,
  parallel = FALSE,
  results_directory = getwd(),
  model_parameter_list = list(),
  algorithm_parameter_list = list(),
  seed = NULL,
  grain_size = 1e+05
)
}
\arguments{
\item{model}{A file containing the model, or a pre-compiled model list.}

\item{number_of_importance_points}{The number of importance points.}

\item{parallel}{(optional) Set to true to perform the importance sampling in parallel, false for serial.}

\item{results_directory}{(optional) The name of the directory to which results will be written.}

\item{model_parameter_list}{(optional) A list containing parameters for the model.}

\item{algorithm_parameter_list}{(optional) A list containing named parameters for the algorithm.}

\item{seed}{(optional) The seed for the random number generator.}

\item{grain_size}{(optional) Sets a minimum chunk size for parallelisation (see https://oneapi-src.github.io/oneTBB/main/tbb_userguide/Controlling_Chunking_os.html).}
}
\value{
Estimate of the marginal likelihood.
}
\description{
Importance sampler.
}
