% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampler.R
\name{importance_sample_cpp}
\alias{importance_sample_cpp}
\title{Importance sampler using cpp functions.}
\usage{
importance_sample_cpp(model, algorithm, messages = FALSE)
}
\arguments{
\item{model}{A model.}

\item{algorithm}{Algorithm details.}

\item{messages}{Set to TRUE to have messages printed to console.}
}
\value{
Importance points.
}
\description{
Importance sampler using cpp functions.
}
