% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampler.R
\name{importance_sample_cpp}
\alias{importance_sample_cpp}
\title{Importance sampler using cpp functions.}
\usage{
importance_sample_cpp(
  model,
  algorithm,
  max_vector_size = 4e+09,
  messages = FALSE
)
}
\arguments{
\item{model}{A model.}

\item{algorithm}{Algorithm details.}

\item{max_vector_size}{If any vector/list has more than this many entries, we need to split the algorithm into multiple batches.}

\item{messages}{Set to TRUE to have messages printed to console.}
}
\value{
Importance points.
}
\description{
Importance sampler using cpp functions.
}
