% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output.R
\name{load_enk_output}
\alias{load_enk_output}
\title{Loading ensemble Kalman output into R memory.}
\usage{
load_enk_output(results.directory, ggsmc = TRUE)
}
\arguments{
\item{ggsmc}{(optional) Output in tidy format for plotting in ggsmc package.}

\item{results_directory}{The folder in which the results are stored.}
}
\value{
A list containing the ensemble members (called particles).
}
\description{
Loading ensemble Kalman output into R memory.
}
