% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output.R
\name{load_mcmc_output}
\alias{load_mcmc_output}
\title{Loading MCMC output into R memory.}
\usage{
load_mcmc_output(results_directory, thinning = 1, for_ggmcmc = TRUE)
}
\arguments{
\item{results_directory}{The folder in which the results are stored.}

\item{thinning}{(optional) The granularity of the thinning of the MCMC output.}

\item{for_ggmcmc}{(optional) Output in tidy format for plotting in ggmcmc package.}
}
\value{
A data frame / tibble containing the MCMC chains.
}
\description{
Loading MCMC output into R memory.
}
