% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output.R
\name{load_mcmc_output}
\alias{load_mcmc_output}
\title{Loading MCMC output into R memory.}
\usage{
load_mcmc_output(results_directory, ggmcmc = TRUE)
}
\arguments{
\item{results_directory}{The folder in which the results are stored.}

\item{ggmcmc}{(optional) Output in tidy format for plotting in ggmcmc package.}
}
\value{
A list containing the MCMC chains.
}
\description{
Loading MCMC output into R memory.
}
