% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output.R
\name{load_smc_output}
\alias{load_smc_output}
\title{Loading SMC output into R memory.}
\usage{
load_smc_output(
  results_directory,
  ggmcmc = FALSE,
  ggsmc = TRUE,
  as.mcmc = FALSE,
  as.enk = FALSE,
  ilike.output = TRUE,
  which.targets = NULL,
  directory_prefix = "ilike",
  external_log_weights = c(0),
  external_target_parameters = "",
  nesting_level = 0,
  factor = 0
)
}
\arguments{
\item{results_directory}{The folder in which the results are stored.}

\item{ggmcmc}{(optional) Output in tidy format for plotting in ggmcmc package.}

\item{ggsmc}{(optional) Output in tidy format for plotting in ggsmc package.}

\item{as.mcmc}{(optional) Output treats particles as different MCMC chains.}

\item{as.enk}{(optional) Output treats particles as an ensemble.}

\item{which.targets}{(optional) The indices of the targets to output (defaults to all).}

\item{directory_prefix}{(optional; for nested output only) The first part of the name of the directory within results_directory that contains the results. (default is "ilike", giving a directory of results_directory/ilike_smc)}

\item{external_log_weights}{(optional; for nested output only) The weights of the importance points external to the current folder. (default is 1, to be used at the top level of nested output)}

\item{external_target_parameters}{(optional; for nested output only) The parameters of the target external to the current folder. (default is "", corresponding to no parameters)}

\item{nesting_level}{(optional; for nested output only) The level of nesting at which to extract points. (default is 0, representing the top level of nested output)}

\item{factor}{(optional; for nested output only) The factor from which to extract points. (default is 0)}

\item{ilike, output}{(optional) Output can be processed by ilike,output package (default is TRUE). Choosing TRUE for this argument is incompatiable with ggmcmc=TRUE, since the two packages require different formatting of the output.}
}
\value{
A list containing the SMC output.
}
\description{
Loading SMC output into R memory.
}
