% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ilike_model.R
\name{parse_ilike_model}
\alias{parse_ilike_model}
\title{Parse .cpp file to give ilike model.}
\usage{
parse_ilike_model(filename)
}
\arguments{
\item{filename}{The name (and path) of the .cpp file containing the model.}
}
\value{
A list containing the model details.
}
\description{
Parse .cpp file to give ilike model.
}
